refactor(runtime): share live detection pipeline construction
