+++ src/runtime/mod.rs
@@ -6,6 +6,8 @@ pub mod logging;
 #[cfg(target_os = "macos")]
 mod macos;
 mod orchestration;
+#[cfg(any(windows, target_os = "linux", target_os = "macos"))]
+mod pipeline;
 pub mod telemetry;
 #[cfg(windows)]
 mod windows;
